Extract tests into dedicated submodules for core modules
